refactor: decoupling data access from web handlers
